Service implementation

+++ src/main/java/so/fine/codingchallenge/service/DealForNonexistingUserException.java
@@ -2,7 +2,7 @@ package so.fine.codingchallenge.service;
 
 public class DealForNonexistingUserException extends RuntimeException {
 
-    public DealForNonexistingUserException(int id) {
+    public DealForNonexistingUserException(long id) {
         super(String.format("Attempting to create a deal for a nonexisting user ID: %d", id));
     }
 }
